add user dashboard wihlist

+++ database/migrations/2024_08_19_075914_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->string('shipping_number')->nullable();
             $table->string('order_number')->unique();
             $table->date('order_date');
-            $table->decimal('total_amount', 10, 2);
+            $table->integer('total_amount');
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->timestamps();
         });
